Added heaps algorithm to obtain permutations
Heap's algorithm is the most efficient way of obtaining all permutations of a list (it does so in the least number of steps). Replaced the previous method of obtaining all permutations.

The notable differences between these two methods are: Heap's algorithm works for a list of length n (whereas the previous hardcoded method only worked for n = 3), and this method produces all permutations before we iterated through them (the previous method returned a single permutation when needed)