Added support for creating potential nxn magic arrays, cleaned up the magic method
Added several functions to create a list of potential nxn magic arrays given the array created by split. Two are hardcoded series of for loops (one for 3x3, one for 4x4), another function iterates over n!^n items and maps the iteration to a potential magic array, and the final is a NOT WORKING recursive method.

Also moved the functions to check whether the potential magic array out of the creation of the magic arrays, which decreased efficiency but increases readability.