Better printouts, a check for sufficient count of squares
Added several printouts to summarise the search.

Added a check to cut down search space: if the splits does not have a sufficient amount of integer squares to create a magic array, it is discarded.

Created a second check to discard tuples within a splits that can't contribute to creating a potential magic array.