Fixed a bug, created mapping function, added file to check whether a split is magic
The bug appended the shape of regular magic arrays at the end rather than the beginning.
Created a function in magic utility that maps each number to each tuple that it appears.
The new file contains methods called in magic dance to determine whether or not the splits is magic or not.