Added a recursive function that returns permutations of the magic dance, plus some helper functions.
The permutations are a superset of the combinations, but aren't much greater. It almost works well.

The first of two helper functions is the same as create_occurrences_dictionary, which handles the case for when the arrays do not contain meta-data pertaining to their shape; both it and its previous function's names have been updates to reflect this.

The other helper function checks whether all elements of an array are also contained within a set. The opposite of is_unique_tuple.