Completed the match algorithm
Good news: it works.
Bad news: it sucks.

So, checking the count occurrences of each value within a potential split is NOT a good check. Turns out, all previous candidates before the count occurrences check pass the test.
In other words, the foundation of this approach is a total no-go.

God dammit; back to the drawing board.