Created some util function, cleaned up is_magic
Created two new functions is mu: the first creates a set whose elements are the collection of (string) tuples. The second takes a single split and checks if it is in the splits; the split is sorted so that order of the elements does not matter.

Also removed all the is_magic_x functions that aren't being used (that is, all but the first one).